Crate roles_integrity

Re-exports

pub use role_claim::*;

pub use roles::*;

pub use admins::*;

pub use progenitors::*;

pub use role_to_assignee::*;

pub use assignee_to_role::*;

pub use pending_unassignments::*;

pub use all_role_claims_deleted_proof::*;

Modules

admins

all_role_claims_deleted_proof

assignee_to_role

linked_devices

pending_unassignments

progenitors

role_claim

role_to_assignee

roles

utils

Structs

AssignRoleInput

Input structure for assigning roles

AssigneeRoleClaimLinkTag

PendingUnassignmentLinkTag

Input structure for unassigning roles

Properties

App properties (progenitor info)

RequestUnassignRoleInput

Input structure for unassigning roles

Enums

AssigneeRoleClaimLinkMode

EntryTypes

LinkTypes

LinkTypes available in the module

UnitEntryTypes

Constants

ADMIN_ROLE

Hardcoded admin role need for creating other roles

ROLE_CLAIM_LINK_TYPE_INDEX

Functions

__num_entry_types

__num_link_types

action_hash

Finding the appropriate action for an action hash

entry_defs

genesis_self_check

validate

Validate function

validate_agent_had_undeleted_role_claim_at_the_time

Validate that agents had acces to the role at the time of the action (an undeleted claim earlier in source chain)

validate_agent_had_undeleted_role_claim_at_the_time_with_zome_index

Validate that agents had acces to the role at the time (with zome index)

validate_agent_joining

Checking the membrane proof of joining agent